requirements.txt
# Soul Room — Dependencies
# Core
PyYAML>=6.0
aiohttp>=3.9.0
requests>=2.31.0
python-dotenv>=1.0.0
cryptography>=41.0.0

# Local model backend
ollama>=0.2.0

# GUI (optional — room works headless without these)
PyQt6>=6.6.0
PyQt6-Qt6>=6.6.0

# Voice (optional)
edge-tts>=6.1.0
SpeechRecognition>=3.10.0

# Database
# sqlite3 is stdlib — no install needed
